% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analyze_turnout}
\alias{analyze_turnout}
\title{Analyze voter turnout}
\usage{
analyze_turnout(data)
}
\arguments{
\item{data}{A data frame containing voter data}
}
\value{
A summary of voter turnout
}
\description{
Analyze voter turnout
}
