% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_R2.R
\name{Bayesian_R2}
\alias{Bayesian_R2}
\title{Marginal and Conditional Bayes R-Square for Bayesian Models}
\usage{
Bayesian_R2(
  model,
  marginal = TRUE,
  conditional = FALSE,
  effects = NULL,
  save = FALSE,
  path = getwd(),
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{model}{A required argument supplying the \code{brmsfit}
or \code{stanreg} model object for which to calculate
Bayes \eqn{R^2}}

\item{marginal}{Logical argument indicating whether to calculate
the unconditional Bayes \eqn{R^2}. Defaults to \code{TRUE}}

\item{conditional}{Logical argument indicating whether to calculate
Bayes \eqn{R^2} conditional on the random effects. Defaults to
\code{FALSE}}

\item{effects}{A one-sided formula specifying the structure of
the random effects for the conditional Bayes \eqn{R^2}. Defaults
to \code{NULL}, which conditions on the full random effects structure.}

\item{save}{Logical argument indicating whether the results should be
saved to the local model file. Defaults to \code{FALSE} and is
currently only supported for \code{brmsfit} objects.}

\item{path}{The base directory for where the updated model object
should be saved. Defaults to the current working directory and is
ignored if save is \code{FALSE}.}

\item{overwrite}{Logical argument indicating whether to overwrite any
existing \code{Bayesian_R2} result saved in model$criteria. Defaults to
\code{FALSE}, which for \code{brmsfit} objects will simply return
the original model object supplied.}

\item{...}{Additional arguments to be passed to either
\code{brms::bayes_R2} or \code{rstanarm::bayes_R2} depending on
the model class.}
}
\value{
The function returns the original model object with the
results stored in the \code{model$criteria$Bayesian_R2}.
}
\description{
This function returns estimates for both the marginal and
conditional Bayes \eqn{R^2} for multilevel models fit with
either \pkg{brms} or \pkg{rstanarm} and stores them
in the original model object.
}
\details{
For additional details on the underlying calls see
\code{?brms::bayes_R2} or \code{?rstanarm::bayes_R2}. A technical
discussion of the Bayes \eqn{R^2} statistic can be found in
\insertCite{Gelman2019}{PoliSciTools}
}
\references{
\insertAllCited{}
}
