% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{clean_political_data}
\alias{clean_political_data}
\title{Clean political science data}
\usage{
clean_political_data(data)
}
\arguments{
\item{data}{A data frame containing political science data}
}
\value{
A cleaned data frame
}
\description{
Clean political science data
}
