% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_IPTW.R
\name{IPTW_Weights}
\alias{IPTW_Weights}
\title{Functions for Constructing Inverse Probability of Treatment Weights for Bayesian Panel Models}
\usage{
IPTW_Weights(
  num_model,
  denom_model,
  method = c("epred", "average"),
  response = NULL,
  id,
  ts,
  trim = FALSE,
  log_weights = FALSE,
  ...
)
}
\arguments{
\item{num_model}{A required argument supplying either a model of class \code{brmsfit}
or an \eqn{m \times n} matrix of posterior expectations where \eqn{m} denotes draws
from the posterior predictive distribution and \eqn{n} is the number of observations
in the data. This argument corresponds to the numerator of the stabilized weights.}

\item{denom_model}{A required argument supplying either a model of class \code{brmsfit}
or an \eqn{m \times n} matrix of posterior expectations where \eqn{m} denotes draws
from the posterior predictive distribution and \eqn{n} is the number of observations
in the data. This argument corresponds to the denominator of the stabilized weights
and must be of same class as \code{num_model}.}

\item{method}{An argument of type character indicating whether \code{num_model} and
\code{denom_model} are \code{matrix} or \code{brmsfit} objects. Defaults to \code{"epred"}
which corresponds to \code{brmsfit} objects but can be set to \code{"average"} when
passing matrices of posterior predictions which allows for the use of model averaged
or stacked posterior predictive distributions.}

\item{response}{A required argument specifying a numeric vector of length \eqn{n}
containing the data for the response used in the treatment propensity model.}

\item{id}{A required argument of length \eqn{n} containing the grouping structure
of the cross-sectional dimension of the data (i.e., states, countries, etc.)}

\item{ts}{A required argument of length \eqn{n} containing the temporal structure
of the time series dimension of the data (i.e., years, months, etc.)}

\item{trim}{A numeric value by which to adjust the predictions before
calculating the weights. Defaults to \code{FALSE} which does not make any
adjustments to the posterior predictions. This can be useful in the case of
predicted treatment probabilities near 0 or 1.}

\item{log_weights}{Whether to return the summarized weights in raw or log transformed
form. This can be useful when the stabilized weights exhibit a high degree of variance.
Defaults to \code{FALSE} which just returns the weights in their raw form. If
\code{TRUE}, summary estimates are transformed by calling \code{log1p}.}

\item{...}{Additional arguments to be passed to \code{brms::posterior_epred}.
Only used if method = "epred" and ignored otherwise.}
}
\value{
The function returns a list containing a matrix of the full posterior
distribution of the inverse probability of treatment weights and by-observation
summary statistics.
}
\description{
Functions for estimating the posterior distribution of stabilized inverse probability of
treatment weights for marginal structural models applied to time series cross-sectional
data.
}
\details{
Following \insertCite{Blackwell2018}{PoliSciTools}, the stabilized inverse probability weights
for the average treatment effect (ATE) of a binary treatment can be expressed as
\deqn{\hat{SW}_{i[t]} = \prod^{t}_{t = 1} \frac{\Pr[X_{i[t]} | \bar{X}_{i[t-1]}, V_i]}{\Pr[X_{i[t]} | \bar{X}_{i[t-1]}, Y_{i[t-1]}, C_{i[t]}, V_{[i]}]}}
}
\references{
\insertAllCited{}
}
