% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analyze_turnout}
\alias{analyze_turnout}
\title{Analyze voter turnout}
\usage{
analyze_turnout(data)
}
\arguments{
\item{data}{A cleaned data frame containing voter data, including a 'turnout' column}
}
\value{
A summary of voter turnout

A summary of voter turnout statistics and turnout by categorical variables
}
\description{
Analyze voter turnout
}
