% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{visualize_party_distribution}
\alias{visualize_party_distribution}
\title{Visualize party distribution}
\usage{
visualize_party_distribution(data)
}
\arguments{
\item{data}{A data frame containing party affiliation data, including a 'party' column}
}
\value{
A ggplot object

A ggplot object showing the distribution of party affiliation
}
\description{
Visualize party distribution
}
